+++ src/gamedata/damage_callback_utils.h
@@ -6,6 +6,27 @@
 #include "statuses.h"
 #include <vector>
 #include <string>
+#include "../party/battle_stats.h"
+#include "../battle/battle_actors.h"
+#include "damage_formulas.h"
+
+struct SrDamageContext_ {
+    std::vector<std::string> attackElements;
+    std::unordered_map<std::string, SrActorStat> attackStats;
+    std::vector<StatusInfliction> statusToInflict; //vector of statuses to inflict
+    std::vector<std::string> toAddStatuses;
+    std::vector<std::string> toToggleStatuses;
+    std::vector<std::string> toRemoveStatuses;
+    std::vector<std::string> wereInflictedStatuses;
+    ActorBattleState attackerState;
+    ActorBattleState targetState;
+    bool useActionDamageLimits;
+    u16 hpDamageLimit;
+    u16 mpDamageLimit;
+    DamageType damageType;
+    DamageFormula damageFormula;
+    HitFormula hitFormula;
+};
 
 bool actorIsDamageImmune(ActorBattleVars battleVars, ActorBattleState srActorState, bool isMagic);
 bool srActorHasStatus(const ActorBattleState& srActorState, std::string statusName);
