+++ src/impl.h
@@ -50,27 +50,8 @@ extern "C" {
 #include "battle/ai_script_engine.h"
 #include "battle/battle_actors.h"
 #include <map>
+#include <string>
 #include <memory>
-
-/*Game Context holds all the registries which contain
-  Resources, in the form of either data or registered callback*/
-struct SrDamageContext_ {
-    std::vector<std::string> attackElements;
-    std::unordered_map<std::string, SrActorStat> attackStats;
-    std::vector<StatusInfliction> statusToInflict; //vector of statuses to inflict
-    std::vector<std::string> toAddStatuses;
-    std::vector<std::string> toToggleStatuses;
-    std::vector<std::string> toRemoveStatuses;
-    std::vector<std::string> wereInflictedStatuses;
-    ActorBattleState attackerState;
-    ActorBattleState targetState;
-    bool useActionDamageLimits;
-    u16 hpDamageLimit;
-    u16 mpDamageLimit;
-    DamageType damageType;
-    DamageFormula damageFormula;
-    HitFormula hitFormula;
-};
 
 typedef struct {
     lua_State*                          L;
